uses buildin buttons for users without keypad
As I don't have the keypad and wanted to test the LNPoS without waiting for it, I have programmed the m5stack buildin button B to increase the amount by +10%.

This might be useful for other developers and even for restaurants that are ok with setting approximate price (as 10% tips are usual in many places).

It also introduces some improvemnts in the UI by showing what each of the buttons does just above it (instead of calling them by names A,B and C).

Reset set the price to 1, while restarting the device sets the price to 0.01 for testing.